move frontend upload logic TaskCompletion --> TaskCompletionBody
