+++ app/src/components/task-completion/TaskUpload.tsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { useState } from 'react';
 
 const FileInputWrapper = styled.div`
   margin-top: 0;
